// extension/src/gen/searchium.ts
export interface TypedRequest {
    /**
     * @generated from protobuf oneof: subtype
     */
    subtype: {
        oneofKind: "registerFileRequest";
        /**
         * @generated from protobuf field: searchium.v1.RegisterFileRequest register_file_request = 10;
         */
        registerFileRequest: RegisterFileRequest;
    } | {
        oneofKind: "unregisterFileRequest";
        /**
         * @generated from protobuf field: searchium.v1.UnregisterFileRequest unregister_file_request = 11;
         */
        unregisterFileRequest: UnregisterFileRequest;
    } | {
        oneofKind: "getFileSystemRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetFileSystemRequest get_file_system_request = 12;
         */
        getFileSystemRequest: GetFileSystemRequest;
    } | {
        oneofKind: "searchFilePathsRequest";
        /**
         * @generated from protobuf field: searchium.v1.SearchFilePathsRequest search_file_paths_request = 13;
         */
        searchFilePathsRequest: SearchFilePathsRequest;
    } | {
        oneofKind: "searchCodeRequest";
        /**
         * @generated from protobuf field: searchium.v1.SearchCodeRequest search_code_request = 14;
         */
        searchCodeRequest: SearchCodeRequest;
    } | {
        oneofKind: "getFileSystemVersionRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetFileSystemVersionRequest get_file_system_version_request = 15;
         */
        getFileSystemVersionRequest: GetFileSystemVersionRequest;
    } | {
        oneofKind: "getFileExtractsRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetFileExtractsRequest get_file_extracts_request = 16;
         */
        getFileExtractsRequest: GetFileExtractsRequest;
    } | {
        oneofKind: "getDirectoryStatisticsRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetDirectoryStatisticsRequest get_directory_statistics_request = 18;
         */
        getDirectoryStatisticsRequest: GetDirectoryStatisticsRequest;
    } | {
        oneofKind: "refreshFilesystemTreeRequest";
        /**
         * @generated from protobuf field: searchium.v1.RefreshFileSystemTreeRequest refresh_filesystem_tree_request = 19;
         */
        refreshFilesystemTreeRequest: RefreshFileSystemTreeRequest;
    } | {
        oneofKind: "getDatabaseStatisticsRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetDatabaseStatisticsRequest get_database_statistics_request = 20;
         */
        getDatabaseStatisticsRequest: GetDatabaseStatisticsRequest;
    } | {
        oneofKind: "pauseIndexingRequest";
        /**
         * @generated from protobuf field: searchium.v1.PauseIndexingRequest pause_indexing_request = 21;
         */
        pauseIndexingRequest: PauseIndexingRequest;
    } | {
        oneofKind: "resumeIndexingRequest";
        /**
         * @generated from protobuf field: searchium.v1.ResumeIndexingRequest resume_indexing_request = 22;
         */
        resumeIndexingRequest: ResumeIndexingRequest;
    } | {
        oneofKind: "getDatabaseDetailsRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetDatabaseDetailsRequest get_database_details_request = 23;
         */
        getDatabaseDetailsRequest: GetDatabaseDetailsRequest;
    } | {
        oneofKind: "getProjectDetailsRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetProjectDetailsRequest get_project_details_request = 24;
         */
        getProjectDetailsRequest: GetProjectDetailsRequest;
    } | {
        oneofKind: "getDirectoryDetailsRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetDirectoryDetailsRequest get_directory_details_request = 25;
         */
        getDirectoryDetailsRequest: GetDirectoryDetailsRequest;
    } | {
        oneofKind: "getFilesystemTreeRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetFilesystemTreeRequest get_filesystem_tree_request = 26;
         */
        getFilesystemTreeRequest: GetFilesystemTreeRequest;
    } | {
        oneofKind: "getDirectoryEntriesRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetDirectoryEntriesRequest get_directory_entries_request = 27;
         */
        getDirectoryEntriesRequest: GetDirectoryEntriesRequest;
    } | {
        oneofKind: "getDirectoryEntriesMultipleRequest";
        /**
         * @generated from protobuf field: searchium.v1.GetDirectoryEntriesMultipleRequest get_directory_entries_multiple_request = 28;
         */
        getDirectoryEntriesMultipleRequest: GetDirectoryEntriesMultipleRequest;
    } | {
        oneofKind: undefined;
    };
}
// ...
export const TypedRequest = new TypedRequest$Type();
// ...
export const RegisterFileRequest = new RegisterFileRequest$Type();
// ...
export const UnregisterFileRequest = new UnregisterFileRequest$Type();
// ...
export const GetFileSystemRequest = new GetFileSystemRequest$Type();
// ...
export const SearchFilePathsRequest = new SearchFilePathsRequest$Type();
// ...
export const SearchCodeRequest = new SearchCodeRequest$Type();
// ...
export const GetFileSystemVersionRequest = new GetFileSystemVersionRequest$Type();
// ...
export const GetFileExtractsRequest = new GetFileExtractsRequest$Type();
// ...
export const GetDirectoryStatisticsRequest = new GetDirectoryStatisticsRequest$Type();
// ...
export const RefreshFileSystemTreeRequest = new RefreshFileSystemTreeRequest$Type();
// ...
export const GetDatabaseStatisticsRequest = new GetDatabaseStatisticsRequest$Type();
// ...
export const PauseIndexingRequest = new PauseIndexingRequest$Type();
// ...
export const ResumeIndexingRequest = new ResumeIndexingRequest$Type();
// ...
export const GetDatabaseDetailsRequest = new GetDatabaseDetailsRequest$Type();
// ...
export const GetProjectDetailsRequest = new GetProjectDetailsRequest$Type();
// ...
export const GetDirectoryDetailsRequest = new GetDirectoryDetailsRequest$Type();
// ...
export const GetFilesystemTreeRequest = new GetFilesystemTreeRequest$Type();
// ...
export const GetDirectoryEntriesRequest = new GetDirectoryEntriesRequest$Type();
// ...
export const GetDirectoryEntriesMultipleRequest = new GetDirectoryEntriesMultipleRequest$Type();
// ...
export const FilePositionSpan = new FilePositionSpan$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SearchParams$Type extends MessageType<SearchParams> {
    constructor() {
        super("searchium.v1.SearchParams", [
            { no: 1, name: "search_string", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "file_path_pattern", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "max_results", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "match_case", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 5, name: "match_whole_word", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 6, name: "include_sym_links", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 7, name: "regex", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 8, name: "use_re2_engine", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<SearchParams>): SearchParams {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.searchString = "";
        message.filePathPattern = "";
        message.maxResults = 0;
        message.matchCase = false;
        message.matchWholeWord = false;
        message.includeSymLinks = false;
        message.regex = false;
        message.useRe2Engine = false;
        if (value !== undefined)
            reflectionMergePartial<SearchParams>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SearchParams): SearchParams {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string search_string */ 1:
                    message.searchString = reader.string();
                    break;
                case /* string file_path_pattern */ 2:
                    message.filePathPattern = reader.string();
                    break;
                case /* int32 max_results */ 3:
                    message.maxResults = reader.int32();
                    break;
                case /* bool match_case */ 4:
                    message.matchCase = reader.bool();
                    break;
                case /* bool match_whole_word */ 5:
                    message.matchWholeWord = reader.bool();
                    break;
                case /* bool include_sym_links */ 6:
                    message.includeSymLinks = reader.bool();
                    break;
                case /* bool regex */ 7:
                    message.regex = reader.bool();
                    break;
                case /* bool use_re2_engine */ 8:
                    message.useRe2Engine = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SearchParams, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string search_string = 1; */
        if (message.searchString !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.searchString);
        /* string file_path_pattern = 2; */
        if (message.filePathPattern !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.filePathPattern);
        /* int32 max_results = 3; */
        if (message.maxResults !== 0)
            writer.tag(3, WireType.Varint).int32(message.maxResults);
        /* bool match_case = 4; */
        if (message.matchCase !== false)
            writer.tag(4, WireType.Varint).bool(message.matchCase);
        /* bool match_whole_word = 5; */
        if (message.matchWholeWord !== false)
            writer.tag(5, WireType.Varint).bool(message.matchWholeWord);
        /* bool include_sym_links = 6; */
        if (message.includeSymLinks !== false)
            writer.tag(6, WireType.Varint).bool(message.includeSymLinks);
        /* bool regex = 7; */
        if (message.regex !== false)
            writer.tag(7, WireType.Varint).bool(message.regex);
        /* bool use_re2_engine = 8; */
        if (message.useRe2Engine !== false)
            writer.tag(8, WireType.Varint).bool(message.useRe2Engine);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message searchium.v1.SearchParams
 */
export const SearchParams = new SearchParams$Type();